Ditch LibQTip for a custom-written tooltip
We need a proper 4-column display to get layout to work properly, but
LibQTip has a couple of bugs and deficiencies that prevent us from using
it with anything but a 2-column display. Unfortunately LibQTip is a
closed project and the maintainer doesn't have the time to actually
review patches, so we can't get the required changes in to LibQTip.

Instead, create our own tooltip API that's designed for just the
functionality we need. It uses frame recycling and should look the same
as LibQTip, but without all of the extra complexity from LibQTip we
weren't using, and without the bugs.

Scrolling support is gone. I haven't decided if this is going to be
restored.